Add support to Rakefile for other kinds of files
This change makes the `rake create` and `rake clean` much more
abstract, replacing a defined array of dotfiles with a Hash to figure
out which files to symlink where. Specifically, this change makes `rake
update` create the proper symlinks for the new file calliope.zsh-theme
created in the previous commit.